% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMCAlite.R
\name{scMCAlite}
\alias{scMCAlite}
\title{Mouse Cell Atlas mapping}
\usage{
scMCAlite(scdata, num = 3)
}
\arguments{
\item{scdata}{matrix-like data input, log-normalized data recommended.}

\item{num}{the number of the most similar cell types returned, default 3.}
}
\value{
a list containing cors_matrix, top_cors, scMCA and scMCA_similarity.
}
\description{
simplized version of origianl scMCA function.
}
